backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
sentence-transformers==2.7.0
faiss-cpu>=1.9.0
requests==2.31.0
python-multipart==0.0.6
huggingface-hub>=0.20.0
transformers>=4.37.0
numpy>=1.24.0
